requirements.txt
opencv-python
numpy
pillow
scikit-image 